FEAT - 계획 삭제 기능 API 추가

+++ src/main/java/com/CapStone/Backend/Repository/InfoRepository.java
@@ -3,6 +3,6 @@ package com.CapStone.Backend.Repository;
 import com.CapStone.Backend.Entity.Info;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InfoRepository extends JpaRepository<Info, Long>  {
+public interface InfoRepository extends JpaRepository<Info, Integer>  {
 
 }
